# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LearnFlow API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/topics/{topic_id}/cards": {
      "get": {
        "tags": [
          "cards"
        ],
        "summary": "List Topic Cards",
        "operationId": "list_topic_cards_api_topics__topic_id__cards_get",
        "parameters": [
          {
            "name": "topic_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Topic Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CardResponse"
                  },
                  "title": "Response List Topic Cards Api Topics  Topic Id  Cards Get"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          }
        }
      }
    },
    "/api/cards/drafts": {
      "get": {
        "tags": [
          "cards"
        ],
        "summary": "List Drafts",
        "description": "Все черновики (status=draft, любой source: ai/mcp).",
        "operationId": "list_drafts_api_cards_drafts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CardResponse"
                  },
                  "type": "array",
                  "title": "Response List Drafts Api Cards Drafts Get"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "Not Implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/cards": {
      "post": {
        "tags": [
          "cards"
        ],
        "summary": "Create Card",
        "description": "Ручное создание → сразу в конец очереди learning.",
        "operationId": "create_card_api_cards_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "Not Implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/cards/{card_id}": {
      "patch": {
        "tags": [
          "cards"
        ],
        "summary": "Update Card",
        "operationId": "update_card_api_cards__card_id__patch",
        "parameters": [
          {
            "name": "card_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Card Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          }
        }
      },
      "delete": {
        "tags": [
          "cards"
        ],
        "summary": "Delete Card",
        "operationId": "delete_card_api_cards__card_id__delete",
        "parameters": [
          {
            "name": "card_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Card Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          }
        }
      }
    },
    "/api/cards/{card_id}/approve": {
      "post": {
        "tags": [
          "cards"
        ],
        "summary": "Approve Card",
        "description": "draft → learning (в конец очереди темы).",
        "operationId": "approve_card_api_cards__card_id__approve_post",
        "parameters": [
          {
            "name": "card_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Card Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftActionResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          }
        }
      }
    },
    "/api/cards/{card_id}/reject": {
      "post": {
        "tags": [
          "cards"
        ],
        "summary": "Reject Card",
        "description": "Удаление черновика.",
        "operationId": "reject_card_api_cards__card_id__reject_post",
        "parameters": [
          {
            "name": "card_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Card Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftActionResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          }
        }
      }
    },
    "/api/learning/state": {
      "get": {
        "tags": [
          "learning"
        ],
        "summary": "Get Learning State",
        "description": "Текущая карточка темы + упражнение + reviews_due + прогресс.",
        "operationId": "get_learning_state_api_learning_state_get",
        "parameters": [
          {
            "name": "topic_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Topic Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningStateResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          }
        }
      }
    },
    "/api/learning/attempt": {
      "post": {
        "tags": [
          "learning"
        ],
        "summary": "Submit Attempt",
        "description": "Проверка ответа, обновление прогресса, следующее упражнение.",
        "operationId": "submit_attempt_api_learning_attempt_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseAttempt"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "Not Implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AttemptResult": {
        "properties": {
          "correct": {
            "type": "boolean",
            "title": "Correct"
          },
          "consecutive_correct": {
            "type": "integer",
            "title": "Consecutive Correct"
          },
          "required_consecutive": {
            "type": "integer",
            "title": "Required Consecutive",
            "default": 3
          },
          "exercise_mastered": {
            "type": "boolean",
            "title": "Exercise Mastered"
          },
          "card_mastered": {
            "type": "boolean",
            "title": "Card Mastered"
          },
          "correct_answer": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Correct Answer",
            "description": "Правильный ответ (возвращается при ошибке)"
          },
          "next_exercise": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Exercise"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "correct",
          "consecutive_correct",
          "exercise_mastered",
          "card_mastered"
        ],
        "title": "AttemptResult"
      },
      "CardCreate": {
        "properties": {
          "topic_id": {
            "type": "integer",
            "title": "Topic Id"
          },
          "type": {
            "$ref": "#/components/schemas/CardType"
          },
          "front_content": {
            "type": "string",
            "minLength": 1,
            "title": "Front Content"
          },
          "back_content": {
            "type": "string",
            "minLength": 1,
            "title": "Back Content"
          }
        },
        "type": "object",
        "required": [
          "topic_id",
          "type",
          "front_content",
          "back_content"
        ],
        "title": "CardCreate",
        "description": "Ручное создание карточки — сразу попадает в конец очереди learning."
      },
      "CardResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "topic_id": {
            "type": "integer",
            "title": "Topic Id"
          },
          "type": {
            "$ref": "#/components/schemas/CardType"
          },
          "status": {
            "$ref": "#/components/schemas/CardStatus"
          },
          "source": {
            "$ref": "#/components/schemas/CardSource"
          },
          "front_content": {
            "type": "string",
            "title": "Front Content"
          },
          "back_content": {
            "type": "string",
            "title": "Back Content"
          },
          "order_index": {
            "type": "integer",
            "title": "Order Index"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "topic_id",
          "type",
          "status",
          "source",
          "front_content",
          "back_content",
          "order_index",
          "created_at"
        ],
        "title": "CardResponse"
      },
      "CardSource": {
        "type": "string",
        "enum": [
          "manual",
          "ai",
          "mcp"
        ],
        "title": "CardSource"
      },
      "CardStatus": {
        "type": "string",
        "enum": [
          "draft",
          "learning",
          "mastered"
        ],
        "title": "CardStatus"
      },
      "CardType": {
        "type": "string",
        "enum": [
          "term",
          "command",
          "procedure"
        ],
        "title": "CardType"
      },
      "CardUpdate": {
        "properties": {
          "front_content": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Front Content"
          },
          "back_content": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Back Content"
          }
        },
        "type": "object",
        "title": "CardUpdate"
      },
      "DraftActionResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/CardStatus"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "message"
        ],
        "title": "DraftActionResponse",
        "description": "Ответ на approve/reject черновика."
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Машиночитаемый код ошибки, напр. 'card_not_found', 'invalid_credentials'"
          }
        },
        "type": "object",
        "required": [
          "detail",
          "code"
        ],
        "title": "ErrorResponse"
      },
      "Exercise": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "payload": {
            "type": "object",
            "title": "Payload"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "payload"
        ],
        "title": "Exercise",
        "description": "Сгенерированное упражнение для текущей карточки.\n\nkind: тип упражнения (см. ExerciseKind в фронтовых типах).\npayload: структура зависит от kind —\n  multiple_choice: {\"options\": [...], \"prompt\": str}\n  text_input: {\"prompt\": str}\n  fill_blank: {\"template\": \"git {{blank}} origin\", \"answer_length\": int}\n  order_steps: {\"shuffled_steps\": [...]}\n  и т.д."
      },
      "ExerciseAttempt": {
        "properties": {
          "card_id": {
            "type": "integer",
            "title": "Card Id"
          },
          "exercise_kind": {
            "type": "string",
            "title": "Exercise Kind"
          },
          "answer": {
            "title": "Answer",
            "description": "str для текстовых, list[str] для order_steps и т.п."
          },
          "used_hint": {
            "type": "boolean",
            "title": "Used Hint",
            "default": false
          },
          "response_time_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Time Ms"
          }
        },
        "type": "object",
        "required": [
          "card_id",
          "exercise_kind",
          "answer"
        ],
        "title": "ExerciseAttempt"
      },
...
      "LearningStateResponse": {
        "properties": {
          "topic_id": {
            "type": "integer",
            "title": "Topic Id"
          },
          "current_card": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CardResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "exercise": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Exercise"
              },
              {
                "type": "null"
              }
            ],
            "description": "Текущее упражнение для current_card"
          },
          "reviews_due": {
            "type": "integer",
            "title": "Reviews Due",
            "default": 0
          },
          "cards_total": {
            "type": "integer",
            "title": "Cards Total",
            "default": 0
          },
          "cards_mastered": {
            "type": "integer",
            "title": "Cards Mastered",
            "default": 0
          },
          "progress_percent": {
            "type": "number",
            "title": "Progress Percent",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "topic_id",
          "current_card"
        ],
        "title": "LearningStateResponse"
      },
...
    }
  }
}